Inventory model
Develop a multi-threaded java program where one thread reads the data from the database say, details of an Item from a mysql table. This thread builds an in-memory object, stores it in a collection. Simultaneously another thread should fetch already created Item objects from this collection and calculate the tax as per rules detailed update the tax value in appropriate Item attribute and store it in a different collection. Finally print out the item details to console as detailed as:-

● All options other than item name can come in any order i.e. after -name you can have -price, -type option. Item type is a mandatory option.
● The system should provide functionality to add more than one items i.e. Once the details of the first item is entered it should display a message saying:
Do you want to enter details of any other item (y/n):
Appropriate behavior is required to be implemented as per user input for this question.
● Make use of java's object oriented capabilities for implementing this business logic.
● Item type can have 3 possible values raw, manufactured and imported.
● Tax rules for the 3 types are as follows:
raw: 12.5% of the item cost
manufactured: 12.5% of the item cost + 2% of (item cost + 12.5% of the item cost)
imported: 10% import duty on item cost + a surcharge (surcharge is: Rs. 5 if the final cost after applying tax & import duty is up to Rs. 100, Rs. 10 if the cost exceeds 100 and up to 200 and 5% of the final cost if it exceeds 200).